initial twitter bootstrap and yuidocs commit
There are going to be a lot of changes happening over the next few days.
I'm working towards organizing the project more by using a gruntjs for
the build process, yuidocs for better documentation and twitter
bootstrap for the style. This commit doesn't have all the features added
back in and may be a bit of a gimp.